Move Kotlin tests to `integTest`
Core does not have the necessary dependencies for its integ test image.
