+++ subprojects/core/src/integTest/groovy/org/gradle/api/tasks/NestedInputIntegrationTest.groovy
@@ -701,90 +701,6 @@ class NestedInputIntegrationTest extends AbstractIntegrationSpec {
         """
     }
 
-    @Requires(TestPrecondition.KOTLIN_SCRIPT)
-    def "implementations in nested Action property in Kotlin build script is tracked"() {
-        setupTaskClassWithNestedAction()
-        def buildFile = file("build.gradle.kts")
-        buildFile << """
-            tasks.create<TaskWithNestedAction>("myTask") {
-                action = Action { writeText("original") }
-            }
-        """
-
-        buildFile.makeOlder()
-
-        when:
-        run 'myTask'
-        then:
-        executedAndNotSkipped(':myTask')
-
-        when:
-        buildFile.text = """
-            tasks.create<TaskWithNestedAction>("myTask") {
-                action = Action { writeText("changed") }
-            }                      
-        """
-        run 'myTask', '--info'
-        then:
-        executedAndNotSkipped(':myTask')
-        file('build/tmp/myTask/output.txt').text == "changed"
-        output.contains "Value of input property 'action.class' has changed for task ':myTask'"
-    }
-
-    @Requires(TestPrecondition.KOTLIN_SCRIPT)
-    def "implementations in nested lambda property in Kotlin build script is tracked"() {
-        file('buildSrc/build.gradle.kts') << """
-            plugins {
-                `kotlin-dsl`
-            }
-        """
-        file("buildSrc/src/main/kotlin/TaskWithNestedLambda.kt") << """
-            import org.gradle.api.DefaultTask
-            import org.gradle.api.tasks.Nested
-            import org.gradle.api.tasks.OutputFile
-            import org.gradle.api.tasks.TaskAction
-            import java.io.File
-            
-            open class TaskWithNestedLambda : DefaultTask() {
-                @get: Nested
-                lateinit var action: (File) -> Unit
-            
-                @get: OutputFile
-                var outputFile: File = File(temporaryDir, "output.txt")
-            
-                @TaskAction
-                fun generate() {
-                    action(outputFile)
-                }
-            }
-        """
-        def buildFile = file("build.gradle.kts")
-        buildFile << """
-            tasks.create<TaskWithNestedLambda>("myTask") {
-                action = { it.writeText("original") }
-            }
-        """
-
-        buildFile.makeOlder()
-
-        when:
-        run 'myTask'
-        then:
-        executedAndNotSkipped(':myTask')
-
-        when:
-        buildFile.text = """
-            tasks.create<TaskWithNestedLambda>("myTask") {
-                action = { it.writeText("changed") }
-            }
-        """
-        run 'myTask', '--info'
-        then:
-        executedAndNotSkipped(':myTask')
-        file('build/tmp/myTask/output.txt').text == "changed"
-        output.contains "Value of input property 'action.class' has changed for task ':myTask'"
-    }
-
     @Requires(TestPrecondition.JDK8_OR_LATER)
     def "implementations in nested property defined by Java 8 lambda is tracked"() {
         setupTaskClassWithNestedAction()
